Separated article controler into service. All services are made with interfaces
